Allow applications to access the loaded function pointers
Without this, an application has to create their own loader for the
Vulkan function pointers if they need to drop down to raw Vulkan
calls. It's not a lot of effort to load them, but it's nicer to just
have them available on the Instance or Device rather than having to
pass an extra parameter around for function pointer access.