Remove seemingly unnecessary `undefined_layout` field?
Upon `Swapchain::new_inner`, `ImageEntry`s are created for each image
with their `undefined_layout` field set to `true`.

Once set, this field is never altered within the vulkano codebase. The
only time it is used following initialisation is when it is read during
calls to `SwapchainAcquireFuture::check_image_access`. For reference,
the case in which I am hitting this method is when attempting to execute
a command buffer that clears a swapchain image using
`clear_color_image`.

If the `undefined_layout` field is `true` and the swapchain image layout
is not `Undefined`, the `check_image_access` function returns an error.
[See here](https://docs.rs/vulkano/0.11.1/src/vulkano/swapchain/swapchain.rs.html#828-835).
This is a problem as the implementation of
`ImageAccess::initial_layout_requirement` and
`ImageAccess::final_layout_requirement` both return
`ImageLayout::PresentSrc` which implies to me that by the time the
command buffer is executed the swapchain image layout *must* be
`PresentSrc` (the call to `add_image_memory_barrier` within
`SyncCommandBufferBuilder::build` [ensures
this](https://docs.rs/vulkano/0.11.1/src/vulkano/command_buffer/synced/base.rs.html#720))
and in turn, this error condition will always be true.

In this PR I remove the unused `undefined_layout` field as well as the
error condition that does not seem to make sense and as a result my
downstream project seems to work fine. However, it is more than likely
that there is some special case that @tomaka intended to cover here and
that there is a better, proper solution that involves checking that
swapchain image layouts are actually initially `PresentSrc` (and not
`Undefined`) and only then setting the `undefined_layout` field to
`false`.

@tomaka could you share your thoughts on this briefly? I'm happy to dive
in and do a proper fix if there is indeed some case you had in mind
here that I'm overlooking.

Eventually, this should close #977.