Fix compiler error in code generated for GLSL struct
I was running into a compiler error generated by the `shader!` macro
while writing a compute shader for a downstream commercial project.
Using `cargo-expand` I was able to find the generated code for the
struct and noticed that the current implementation would always generate
an invalid type. *`{}` is never a valid length for a fixed size array of
bytes.*

It seemed to me that `{}` was used within the `quote!` macro
accidentally rather than `#diff` as this particular field is used to
specify a padding amount for matching the memory layout of the struct.
By changing the fixed size array len to use `#diff` I was able to
complete compilation successfully and my code runs as expected.

@tomaka can you confirm that this was likely a bug and that this is the
correct fix? I want to check that this was not an intentional compile
error pending future work or a better error message.