instance: store ray tracing properties
Chain `PhysicalDeviceRayTracingPropertiesKHR` and
`PhysicalDeviceRayTracingPropertiesNV` in `PhysicalDeviceProperties2KHR`
in order to retrieve the same data from two different extensions if they
are available.
In the case both are unavailable the values are 0. In the case of one or
both extensions being supported, grab the maximum, overwriting the 0s in
the case of one of both unsupported and reconciling both properties if they
are available.
One exception is the property `shaderGroupHandleCaptureReplaySize` which
is not available in the nvidia ray tracing extension.