Create build.rs
Thanks for this crate! It's awesome!

This PR just adds a little tweak to make it easier to link binaries to libvulkan. This `build.rs` is useful on platforms where `LD_LIBRARY_PATH` (or equivalent) is not set system-wide, so that the rpath of executables made using vulkano contains libvulkan. The `cfg!` matches the conditions in `vulkano/vulkano/src/instance/loader.rs`.