Fix sync bug in same image blit/copy commands
The bug was introduced by #1782.
When we implemented "same image blit/copy" feature,
we added only `one` image to the reserved resources for the command if
the same image was used as `source` and `destination`.

The bug happens if we passed the same image, but with different
mip level and/or array layer. In that case, we will only add the
`source` resource and not the `destination`, and that would result in
failing to know when to flush the pipeline barrier.