Implement is_layout_initialized for ImmutableImage
The default implementation always returns `false`. However, in
`ImmutableImage`'s case, by the time we are doing operations with
`ImmutableImage`, the layout should have already been initialized by
`ImmutableImageInitialization`.

Without this function, vulkano thinks the image referenced by the
`ImmutableImage` isn't initialized, and will insert a pipeline
transition from Undefined to `self.layout`, which isn't incorrect; the
command buffer created to initialize the image had a
`final_layout_requirement` of the final layout, and thus had already
transitioned the layout to the correct layout.

See, for example, the discussion in #1578.